+++ pylabianca/stats.py
@@ -109,9 +109,8 @@ def permutation_test(*arrays, paired=False, n_perm=1_000, progress=False,
             return stat
 
 
-# TODO: auto-infer paired from xarray
 def cluster_based_test(frate, compare='image', cluster_entry_pval=0.05,
-                       paired=False, stat_fun=None, n_permutations=1_000,
+                       paired=None, stat_fun=None, n_permutations=1_000,
                        n_stat_permutations=0, tail=None, progress=True,
                        return_clusters=None):
     '''Perform cluster-based tests on firing rate data.
@@ -135,20 +134,24 @@ def cluster_based_test(frate, compare='image', cluster_entry_pval=0.05,
         categories to test selectivity for.
     cluster_entry_pval : float
         p value used as a cluster-entry threshold. The default is ``0.05``.
-    paired : bool
+    paired : bool | None
         Whether a paired (repeated measures) or unpaired test should be used.
+        Defaults to ``None`` which infers the type of test from the structure
+        of the DataArray.
     return_clusters : bool
         Whether to return a ``borsar.Clusters`` object instead of a
         ``(stat, clusters, pval)`` tuple.
 
     Returns
     -------
-    stats : numpy.ndarray
-        Anova F statistics for every time point.
+    stats : numpy.ndarray | borsar.Clusters
+        Value of the test statistic for every time point. If
+        ``return_clusters=True``, then a ``borsar.Clusters`` object is
+        returned instead.
     clusters : list of numpy.ndarray
-        List of cluster memberships.
+        List of cluster memberships. Only returned if ``return_clusters=False``.
     pval : numpy.ndarray
-        List of p values from anova.
+        List of p values from anova. Only returned if ``return_clusters=False``.
     '''
     from borsar.cluster import permutation_cluster_test_array
 
@@ -160,8 +163,7 @@ def cluster_based_test(frate, compare='image', cluster_entry_pval=0.05,
                       ' old behavior use `return_clusters=False`.',
                       FutureWarning)
 
-    # TODO: check if theres is a condition dimension (if so -> paired)
-    arrays = [arr.values for _, arr in frate.groupby(compare)]
+    arrays, levels, dimnames, paired = _prepare_arrays(frate, compare, paired)
 
     if tail is None:
         n_groups = len(arrays)
@@ -175,35 +177,72 @@ def cluster_based_test(frate, compare='image', cluster_entry_pval=0.05,
 
     if return_clusters:
         from borsar.cluster.obj import Clusters
-
-        dimnames, dimcoords = _infer_cluster_coords(frate, compare)
-        return Clusters(stat, clusters=clusters, pvals=pval,
-                        dimnames=dimnames, dimcoords=dimcoords)
+        dimcoords = [frate.coords[dimname].values for dimname in dimnames]
+        desc = { 'compare': compare, 'levels': levels, 'paired': paired,
+                'tail': tail, 'n_permutations': n_permutations,
+                'n_stat_permutations': n_stat_permutations,
+                'cluster_entry_p_threshold': cluster_entry_pval}
+        return Clusters(
+            stat, clusters=clusters, pvals=pval,
+            dimnames=dimnames, dimcoords=dimcoords, description=desc
+        )
 
     return stat, clusters, pval
 
 
-# TODO: this might later need checking against other xarray helpers
-#       to de-duplicate (and some code present in selectivity module
-#       for xarray construction!)
-def _infer_cluster_coords(xarr, compare):
-    dimnames = None
+# TODO: looking for reduced dimensions, transposing, etc. should be checked
+#       against other xarray helpers to de-duplicate (some code present in
+#       selectivity module for xarray construction!)
+def _prepare_arrays(frate, compare, paired):
+    '''Extract and convert arrays format expected by the cluster based test.
 
-    # infer reduced dimension from compare coordinate / dimension
-    if compare in xarr.coords:
-        compare_dims = xarr.coords[compare].dims
-        assert len(compare_dims) == 1
-        compare = compare_dims[0]
+    Additionally: a) returns inferred dimnames for the stats array after
+    conducting the test and b) auto-detects if paired test should be conducted
+    based on the DataArray structure.'''
 
-    n_orig_dims = len(xarr.dims)
-    dimnames = [dim for dim in xarr.dims if dim != compare]
+    if compare not in frate.coords:
+        raise ValueError(f'``compare`` ("{compare}") was not found in'
+                         ' the provided DataArray coordinates.')
 
-    if len(dimnames) == n_orig_dims:
-        raise RuntimeError('Could not find the reduced dimension.')
+    # extract the arrays grouped by the comapre condition
+    arrays, levels = list(), list()
+    for lvl, arr in frate.groupby(compare):
+        arrays.append(arr.values)
+        levels.append(lvl)
 
-    coords = [xarr.coords[dim_name].values for dim_name in dimnames]
+    # prepare arrays for analysis
+    drop_dims = list()
+    if compare in frate.dims:
+        drop_dims.append(compare)
+        dim_idx = frate.dims.index(compare)
+        arrays = [np.squeeze(arr, axis=dim_idx) for arr in arrays]
 
-    return dimnames, coords
+        # first dimension (observations) should be reduced
+        # TODO: auto-detect observation dimension and make sure it is first
+        first_dim = 0 + int(dim_idx == 0)
+        drop_dims.append(frate.dims[first_dim])
+
+        if paired is None:
+            paired = True
+    else:
+        reduced_dim = frate.coords[compare].dims
+        assert len(reduced_dim) == 1
+        reduced_dim = reduced_dim[0]
+        reduced_dim_idx = frate.dims.index(reduced_dim)
+        drop_dims.append(reduced_dim)
+
+        if not reduced_dim_idx == 0:
+            dim_ord = np.arange(arrays[0].ndim)
+            dim_ord = np.delete(dim_ord, reduced_dim_idx)
+            dim_ord = np.concatenate([[reduced_dim_idx], dim_ord])
+            arrays = [arr.transpose(*dim_ord) for arr in arrays]
+
+        if paired is None:
+            paired = False
+
+    dimnames = [dimname for dimname in frate.dims
+                if dimname not in drop_dims]
+    return arrays, levels, dimnames, paired
 
 
 # ENH: move to sarna/borsar sometime
@@ -472,9 +511,8 @@ def _catch_common_percentile_errors(percentile, dist, tail):
     # the user wanted percentile of 5, and not 0.05)
     if percentile < 1:
         import warnings
-        per_text = f'{percentile:.2f} %'
-        warnings.warn('Percentile is very low ({per_text}). Remember that it '
-                      'is a percentile, not a fraction.')
+        warnings.warn(f'Percentile is very low ({percentile:.2f} %). Remember'
+                      ' that it is a percentile, not a fraction.')
 
     # additionally - if tail is 'both' (the default), check if the distribution
     # indeed contain positive and negative values - if not, warn the user